pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "virtual-coffee-platform"
version = "0.1.0"
description = "Virtual Coffee Platform for scheduling and managing virtual coffee meetings"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.1",
    "pydantic>=1.10.7",
    "boto3>=1.26.115",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-asyncio",
    "mypy",
    "ruff",
    "bandit",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
# Enable Pyflakes (`F`), isort (`I`), and other recommended rules
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "B", "COM", "C4", "DTZ", "T10", "ISC", "ICN", "PIE", "PYI", "PT", "Q", "SIM", "TID", "ARG", "ERA", "PD", "PGH", "PL", "TRY", "UP", "RUF"]
ignore = []

# Exclude a variety of commonly ignored directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.ruff.isort]
known-first-party = ["models", "repositories", "services", "api"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]

[tool.pytest.ini_options]
testpaths = ["backend/api/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
